Refactor the moj sortable table component into it's own view utils method.
In an upcoming commit this code gets reused exactly as is. Pulling it out
now to make the next review smaller.
