+++ server/utils/viewUtils.ts
@@ -1,6 +1,9 @@
 import * as nunjucks from 'nunjucks'
 import { SummaryListItem } from './summaryList'
-import { ErrorSummaryArgs, SummaryListArgs } from './govukFrontendTypes'
+import { ErrorSummaryArgs, SummaryListArgs, TableArgs } from './govukFrontendTypes'
+
+export type SortableTableHeaders = { text: string; sort: 'ascending' | 'descending' | 'none' }[]
+export type SortableTableRow = { text: string; sortValue: string | null; href: string | null }[]
 
 export default class ViewUtils {
   static escape(val: string): string {
@@ -49,6 +52,36 @@ export default class ViewUtils {
             return { text: item.lines[0] || '' }
           })(),
         }
+      }),
+    }
+  }
+
+  static sortableTable(headers: SortableTableHeaders, rows: SortableTableRow[]): TableArgs {
+    return {
+      head: headers.map(heading => {
+        return {
+          text: heading.text,
+          attributes: {
+            'aria-sort': heading.sort,
+          },
+        }
+      }),
+      rows: rows.map(row => {
+        return row.map(cell => {
+          const result: Record<string, unknown> = {}
+
+          if (cell.sortValue !== null) {
+            result.attributes = { 'data-sort-value': cell.sortValue }
+          }
+
+          if (cell.href === null) {
+            result.text = cell.text
+          } else {
+            result.html = `<a href="${cell.href}" class="govuk-link">${ViewUtils.escape(cell.text)}</a>`
+          }
+
+          return result
+        })
       }),
     }
   }
